+++ sources/controles/VooControle.cpp
@@ -22,12 +22,12 @@ std::list<Voo *> VooControle::obterTodosOsVoos()
     return this->getVooServico().obterTodosOsVoos();
 }
 
-Voo *VooControle::obterVooPorId(const unsigned long id)
+Voo *VooControle::obterVooPorId(const unsigned int id)
 {
     return this->getVooServico().obterVooPorId(id);
 }
 
-bool VooControle::atualizarVooPorId(unsigned long id, Voo novoVoo)
+bool VooControle::atualizarVooPorId(unsigned int id, Voo novoVoo)
 {
     return this->getVooServico().atualizarVooPorId(id, novoVoo);
 }
@@ -37,7 +37,7 @@ bool VooControle::cadastrarVoo(Voo *voo)
     return this->getVooServico().cadastrarVoo(voo);
 }
 
-bool VooControle::excluirVooPorId(const unsigned long id)
+bool VooControle::excluirVooPorId(const unsigned int id)
 {
     return this->getVooServico().excluirVooPorId(id);
 }
